Tie myself up in knots with recursion
Game doesn't return white moves in this iteration, some tests are
failing and I was highly confused by the end of the day. Last thing I
was working on was attempting to build a game tree using the leaf
generator and recursion. Seems leaves were getting all piled onto the
root node instead of cascading down.