Add player_to_move
Having discovered that nodes should be mimimizer or maximizer based on
which player should move next, I've added the player_to_move property
to MinimaxNode and GoNode so that this can be used in the minimax with
alpha-beta pruning algorithm later. If possible I'll remove the player
property from nodes later.