Working towards fixing breadth issue
Board was only building depth along left path and not breadth. I have
begun working on a fix, but in the mean time built in a new bug where
the built tree ends up with no children at all at the root node after
two moves on the board (when the first move is 0, 0). I think making the
size of the board and the win state configurable will help to debug as
manual testing on 9x9 board with 5 to win is slowing things down.